+++ model/task.go
@@ -1,8 +1,9 @@
 package model
 
+// Task ....
 type Task struct {
 	ID          int64  `db:"id"`
 	Description string `db:"description"`
-	Price       int64  `db:"price"`
+	Price       int32  `db:"price"`
 	Status      string `db:"status"`
 }
